requirements.txt
httpx
python-dotenv
python-telegram-bot
pandas
numpy
pyyaml
pytest
pytest-asyncio
